requirements/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
pandas==1.3.4
PyJWT==2.3.0
pyodbc==4.0.32
scipy==1.7.2
scikit-learn==1.0.1
python_dotenv==0.19.2
portion==2.2.0
celery==5.2.6
redis==4.2.2